requirements.txt
streamlit>=1.24.0
openai>=1.35.0
python-docx>=0.8.11
reportlab>=4.0.0
faiss-cpu>=1.7.4
pypdf>=3.16.0
python-dotenv>=1.0.0
